requirements.txt
# KITT Backend - Requirements
# Version refactorisée et modulaire

# Framework Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# IA et ML
openai==1.3.5
anthropic==0.7.1
sentence-transformers>=5.1.2  # Pour similarité sémantique (embeddings)

# Audio
soundfile==0.12.1
numpy==1.24.3

# Validation de données
pydantic==2.5.0
pydantic-settings==2.1.0

# Configuration
python-dotenv==1.0.0

# Logging (optionnel mais recommandé)
python-json-logger==2.0.7
